seperate ExploreCardViewController, clearer reusable enqueue/dequeue
