+++ Wikipedia/Code/ExploreViewController.swift
@@ -1,119 +1,6 @@
 import UIKit
 import WMF
 
-class ExploreCardViewController: ColumnarCollectionViewController, CardContent {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        register(AnnouncementCollectionViewCell.self, forCellWithReuseIdentifier: "AnnouncementCollectionViewCell", addPlaceholder: true)
-        register(ArticleRightAlignedImageCollectionViewCell.self, forCellWithReuseIdentifier: "ArticleRightAlignedImageCollectionViewCell", addPlaceholder: true)
-        register(RankedArticleCollectionViewCell.self, forCellWithReuseIdentifier: "RankedArticleCollectionViewCell", addPlaceholder: true)
-        register(ArticleFullWidthImageCollectionViewCell.self, forCellWithReuseIdentifier: "ArticleFullWidthImageCollectionViewCell", addPlaceholder: true)
-        register(NewsCollectionViewCell.self, forCellWithReuseIdentifier: "NewsCollectionViewCell", addPlaceholder: true)
-        register(OnThisDayExploreCollectionViewCell.self, forCellWithReuseIdentifier: "OnThisDayExploreCollectionViewCell", addPlaceholder: true)
-        register(WMFNearbyArticleCollectionViewCell.wmf_classNib(), forCellWithReuseIdentifier: WMFNearbyArticleCollectionViewCell.wmf_nibName())
-        register(WMFPicOfTheDayCollectionViewCell.wmf_classNib(), forCellWithReuseIdentifier: WMFPicOfTheDayCollectionViewCell.wmf_nibName())
-    }
-    
-    var dataStore: MWKDataStore!
-    
-    public var contentGroup: WMFContentGroup! {
-        didSet {
-            collectionView.reloadData()
-        }
-    }
-    
-    override func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    
-    override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard let preview = contentGroup.contentPreview as? [Any] else {
-            return 1
-        }
-        let countOfFeedContent = preview.count
-        switch contentGroup.contentGroupKind {
-        case .news:
-            return 1
-        case .onThisDay:
-            return 1
-        case .relatedPages:
-            return min(countOfFeedContent, Int(contentGroup.maxNumberOfCells()) + 1)
-        default:
-            return min(countOfFeedContent, Int(contentGroup.maxNumberOfCells()))
-        }
-    }
-    
-    private func resuseIdentifierAt(_ indexPath: IndexPath) -> String {
-        return "ArticleRightAlignedImageCollectionViewCell"
-    }
-    
-    private func articleURL(forItemAt indexPath: IndexPath) -> URL? {
-        let displayType = contentGroup.displayTypeForItem(at: indexPath.row)
-        var index = indexPath.row
-        switch displayType {
-        case .relatedPagesSourceArticle:
-            return contentGroup.articleURL
-        case .relatedPages:
-            index = indexPath.row - 1
-        case .ranked:
-            guard let content = contentGroup.contentPreview as? [WMFFeedTopReadArticlePreview], content.count > indexPath.row else {
-                return nil
-            }
-            return content[indexPath.row].articleURL
-        default:
-            break
-        }
-        
-        if let contentURL = contentGroup.contentPreview as? URL {
-            return contentURL
-        }
-        
-        guard let content = contentGroup.contentPreview as? [URL], content.count > index else {
-            return nil
-        }
-        
-        return content[index]
-    }
-    
-    private func configure(cell: UICollectionViewCell, forItemAt indexPath: IndexPath, layoutOnly: Bool) {
-        guard let cell = cell as? ArticleCollectionViewCell else {
-            return
-        }
-        guard let articleURL = articleURL(forItemAt: indexPath), let article = dataStore.fetchArticle(with: articleURL) else {
-            return
-        }
-        cell.configure(article: article, displayType: WMFFeedDisplayType.page, index: 0, count: 0, theme: theme, layoutOnly: layoutOnly)
-    }
-    
-    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: resuseIdentifierAt(indexPath), for: indexPath)
-        configure(cell: cell, forItemAt: indexPath, layoutOnly: false)
-        return cell
-    }
-
-    override func collectionView(_ collectionView: UICollectionView, estimatedHeightForItemAt indexPath: IndexPath, forColumnWidth columnWidth: CGFloat) -> WMFLayoutEstimate {
-        var estimate = WMFLayoutEstimate(precalculated: false, height: 100)
-        guard let placeholderCell = placeholder(forCellWithReuseIdentifier: resuseIdentifierAt(indexPath)) as? CollectionViewCell else {
-            return estimate
-        }
-        configure(cell: placeholderCell, forItemAt: indexPath, layoutOnly: true)
-        estimate.height = placeholderCell.sizeThatFits(CGSize(width: columnWidth, height: UIViewNoIntrinsicMetric), apply: false).height
-        estimate.precalculated = true
-        return estimate
-    }
-
-    override func metrics(withBoundsSize size: CGSize, readableWidth: CGFloat) -> WMFCVLMetrics {
-        return WMFCVLMetrics.singleColumnMetrics(withBoundsSize: size, readableWidth: readableWidth, interItemSpacing: 0, interSectionSpacing: 0)
-    }
-
-    func sizeThatFits(_ size: CGSize, apply: Bool) -> CGSize {
-        self.layout.invalidateLayout()
-        self.layout.prepare()
-        return self.layout.collectionViewContentSize
-    }
-    
-}
 
 class ExploreViewController: ColumnarCollectionViewController {
     fileprivate let cellReuseIdentifier = "org.wikimedia.explore.card.cell"
@@ -124,7 +11,7 @@ class ExploreViewController: ColumnarCollectionViewController {
     }
     
     private var cardViewControllers: [IndexPath: ExploreCardViewController] = [:]
-    private var reusableCardViewControllers: Set<ExploreCardViewController> = []
+    private var reusableCardViewControllers: [ExploreCardViewController] = []
 
     
     private var fetchedResultsController: NSFetchedResultsController<WMFContentGroup>!
@@ -168,43 +55,59 @@ class ExploreViewController: ColumnarCollectionViewController {
             return maybeCell
         }
         cell.apply(theme: theme)
-        configure(cell: cell, forItemAt: indexPath, layoutOnly: false)
+        let width = self.layout.layoutAttributesForItem(at: indexPath)?.bounds.size.width ?? 1
+        configure(cell: cell, forItemAt: indexPath, width: width, layoutOnly: false)
         return cell
     }
     
-    func configure(cell: ExploreCardCollectionViewCell, forItemAt indexPath: IndexPath, layoutOnly: Bool) {
-        let group = fetchedResultsController.object(at: indexPath)
-        var cardVC = cardViewControllers[indexPath]
-        if cardVC == nil {
-            cardVC = reusableCardViewControllers.first
-            if cardVC == nil {
-                let newCardVC = ExploreCardViewController()
-                newCardVC.dataStore = dataStore
-                newCardVC.contentGroup = group
-                newCardVC.view.isHidden = true
-                addChildViewController(newCardVC)
-                view.addSubview(newCardVC.view)
-                didMove(toParentViewController: self)
-                cardVC = newCardVC
-            } else {
-                cardVC?.contentGroup = group
-            }
-            if let cardVC = cardVC, !layoutOnly {
-                cardViewControllers[indexPath] = cardVC
-                reusableCardViewControllers.remove(cardVC)
-            }
+    func dequeueReusableCardViewController() -> ExploreCardViewController {
+        if let cardVC = reusableCardViewControllers.last {
+            reusableCardViewControllers.removeLast()
+            return cardVC
         }
-        cell.cardContent = cardVC
+        
+        let cardVC = ExploreCardViewController()
+        cardVC.dataStore = dataStore
+        cardVC.view.isHidden = true
+        cardVC.view.autoresizingMask = []
+        cardVC.view.translatesAutoresizingMaskIntoConstraints = false
+        addChildViewController(cardVC)
+        view.addSubview(cardVC.view)
+        cardVC.didMove(toParentViewController: self)
+        return cardVC
+    }
+    
+    func enqueueReusableCardViewController(_ cardVC: ExploreCardViewController) {
+        cardVC.view.removeFromSuperview()
+        cardVC.view.isHidden = true
+        view.addSubview(cardVC.view)
+        reusableCardViewControllers.append(cardVC)
+    }
+    
+    func configure(cell: ExploreCardCollectionViewCell, forItemAt indexPath: IndexPath, width: CGFloat, layoutOnly: Bool) {
+        let group = fetchedResultsController.object(at: indexPath)
+        let cardVC = dequeueReusableCardViewController()
+        assert(cardVC.view.superview === view)
+        cardVC.view.frame = CGRect(origin: .zero, size: CGSize(width: width, height: 1))
+        cardVC.view.layoutIfNeeded()
+        cardVC.contentGroup = group
+        cell.cardContentSize = cardVC.collectionView.contentSize
+        if layoutOnly {
+            enqueueReusableCardViewController(cardVC)
+        } else {
+            cell.cardContent = cardVC
+        }
         cell.titleLabel.text = group.headerTitle()
         cell.subtitleLabel.text = group.headerSubTitle()
         cell.footerButton.setTitle(group.moreTitle(), for: .normal)
     }
     
     func collectionView(_ collectionView: UICollectionView, didEndDisplaying cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
-        guard let vc = cardViewControllers.removeValue(forKey: indexPath) else {
+        guard let cell = cell as? ExploreCardCollectionViewCell, let vc = cell.cardContent as? ExploreCardViewController else {
             return
         }
-        reusableCardViewControllers.insert(vc)
+        cell.cardContent = nil
+        enqueueReusableCardViewController(vc)
     }
     
 }
@@ -224,7 +127,7 @@ extension ExploreViewController {
             return estimate
         }
         placeholderCell.prepareForReuse()
-        configure(cell: placeholderCell, forItemAt: indexPath, layoutOnly: true)
+        configure(cell: placeholderCell, forItemAt: indexPath, width: columnWidth, layoutOnly: true)
         estimate.height = placeholderCell.sizeThatFits(CGSize(width: columnWidth, height: UIViewNoIntrinsicMetric), apply: false).height
         estimate.precalculated = true
         return estimate
